And here we are, less than an hour and a half till due. The project is incomplete, but so close to a finish.
My brain is ground down after many simple issues going unnoticed.

Hopefully this is judged based not on the completeness, but on the quality of that which is complete.

All I have to showcase are the up and down motions of the blocks and the well-written classes supporting them.

Right and left do not properly traverse the arrays in order to reposition all of the Blocks and still result in errors.

It is time to rest myself prior to the oncoming exam.